remove matchW file
