+++ src/index.ts
@@ -4,10 +4,4 @@
  */
 import * as match from './match'
 
-/**
- * @since 2.0.2
- *
- */
-import * as matchW from './matchW'
-
-export { match, matchW }
+export { match }
